# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexusTale API
  version: "1.0.0"
  description: >
    REST API for NexusTale — a sci-fi/fantasy novel-writing platform with
    Git-backed versioning, world wiki, and AI co-author support.
servers:
  - url: /api/v1
    description: Current API prefix (served by nginx reverse-proxy in all environments)
security: []  # default unauthenticated; individual operations override via bearerAuth
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable error shapes ───────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Validation or binding error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message: { type: string }
        detail:  { type: string }
...
    CreateProjectRequest:
      type: object
      required: [title]
      properties:
        title:       { type: string, minLength: 1, maxLength: 200 }
        description: { type: string }
        genres:      { type: array, items: { type: string } }
...
    ProjectResponse:
      type: object
      required: [id, owner_id, title, description, genres, archived, phase, auto_tag_enabled, created_at, updated_at]
      properties:
        id:               { type: string, format: uuid }
        owner_id:         { type: string, format: uuid }
        title:            { type: string }
        description:      { type: string }
        genres:           { type: array, items: { type: string } }
        archived:         { type: boolean }
        phase:            { type: string }
        auto_tag_enabled: { type: boolean }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
...
    CreateActRequest:
      type: object
      required: [title]
      properties:
        title:      { type: string, minLength: 1, maxLength: 200 }
        summary:    { type: string }
        sort_order: { type: integer }
...
    ActResponse:
      type: object
      required: [id, project_id, title, summary, sort_order, created_at, updated_at]
      properties:
        id:         { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        title:      { type: string }
        summary:    { type: string }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ── Chapters ──────────────────────────────────────────────────────────────
    CreateChapterRequest:
      type: object
      required: [title]
      properties:
        title:      { type: string, minLength: 1, maxLength: 200 }
        summary:    { type: string }
        sort_order: { type: integer }
...
    ChapterResponse:
      type: object
      required: [id, project_id, act_id, title, summary, sort_order, created_at, updated_at]
      properties:
        id:         { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        act_id:     { type: string, format: uuid }
        title:      { type: string }
        summary:    { type: string }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ── Scenes ────────────────────────────────────────────────────────────────
    CreateSceneRequest:
      type: object
      required: [title]
      properties:
        title:      { type: string, minLength: 1, maxLength: 200 }
        content:    { type: string }
        pov:        { type: string }
        tense:      { type: string }
        tags:       { type: array, items: { type: string } }
        summary:    { type: string }
        sort_order: { type: integer }
...
    SceneResponse:
      type: object
      required: [id, chapter_id, title, content, sort_order, word_count, created_at, updated_at]
      properties:
        id:            { type: string, format: uuid }
        chapter_id:    { type: string, format: uuid }
        title:         { type: string }
        content:       { type: string }
        pov:           { type: string }
        tense:         { type: string }
        tags:          { type: array, items: { type: string } }
        summary:       { type: string }
        summary_stale: { type: boolean }
        sort_order:    { type: integer }
        word_count:    { type: integer, description: "Server-computed word count; updated on every content save." }
        created_at:    { type: string, format: date-time }
        updated_at:    { type: string, format: date-time }

    # ── API Keys ──────────────────────────────────────────────────────────────
...
    ChronicleRequest:
      type: object
      required: [note]
      properties:
        note: { type: string, minLength: 1, maxLength: 500 }
    ChronicleEntry:
      type: object
      required: [sha, short_sha, note, author, created_at]
      properties:
        sha:        { type: string }
        short_sha:  { type: string }
        note:       { type: string }
        author:     { type: string }
        created_at: { type: string, format: date-time }

    # Chronicle returns 201 with ChronicleEntry on success, or 200 with this
    # shape when there was nothing new to commit.
    NothingToChronicle:
      type: object
      required: [message]
      properties:
        message:        { type: string, example: "nothing to chronicle" }
        last_chronicle: { $ref: '#/components/schemas/ChronicleEntry', nullable: true }
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: List all projects owned by the authenticated user
      tags: [projects]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of projects.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProjectResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createProject
      summary: Create a new project
      tags: [projects]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectRequest' }
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /projects/{projectId}/acts:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listActs
      summary: List all acts in a project
      tags: [acts]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of acts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ActResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createAct
      summary: Create a new act
      tags: [acts]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateActRequest' }
      responses:
        "201":
          description: Act created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /projects/{projectId}/acts/{actId}/chapters:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: actId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listChapters
      summary: List all chapters in an act
      tags: [chapters]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of chapters.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChapterResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createChapter
      summary: Create a new chapter in an act
      tags: [chapters]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateChapterRequest' }
      responses:
        "201":
          description: Chapter created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChapterResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /chapters/{chapterId}/scenes:
    parameters:
      - name: chapterId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listScenes
      summary: List all scenes in a chapter
      tags: [scenes]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of scenes.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SceneResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createScene
      summary: Create a new scene
      tags: [scenes]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSceneRequest' }
      responses:
        "201":
          description: Scene created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SceneResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /projects/{projectId}/git/chronicle:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: chronicle
      summary: >
        Snapshot all scene content and create a commit (Chronicle) on the
        current Timeline. Returns 201 with the entry on success, or 200 with
        a "nothing to chronicle" message when nothing changed.
      tags: [git]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChronicleRequest' }
      responses:
        "201":
          description: Chronicle created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChronicleEntry' }
        "200":
          description: Nothing to chronicle.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NothingToChronicle' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
...